Fix Next 15 dynamic route param typing with server wrappers
